% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{myncurve}
\alias{myncurve}
\title{mycurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Upper bound of the lower tail}
}
\value{
Displays the curve, shaded area between the curve and x axis from -∞ to x=a, and calculate the area (probability, P(X<=a)) which is released to the command-line in a list.
}
\description{
mycurve
}
