% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Makes a Sample Plot}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample size}

\item{time}{}
}
\value{
A barplot
}
\description{
This function creates barplots of 30 iterations
}
