% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{N variable}

\item{gamma}{gamma variable}

\item{p}{p variable}
}
\value{
plots and lists
}
\description{
Title ntickets
}
